新增Login相关

+++ app/src/main/java/com/king/easychat/netty/codec/PacketHelper.kt
@@ -27,7 +27,7 @@ object PacketHelper {
             .writeByte(packet.version)//版本
             .writeByte(JSON)//json
             .writeByte(packet.messageType())
-            .writeByte(bytes.size)
+            .writeInt(bytes.size)
             .writeBytes(bytes)
 
     }
